events partly modified, 2 methods added to billingServerSecureinterface, account for auctionServer added, events should be reported to analytics-server(if one is online)...

+++ dslab/src/analyticsServer/event/AuctionEvent.java
@@ -2,17 +2,17 @@ package analyticsServer.event;
 
 
 public class AuctionEvent extends Event {
-	private int auctionID;
+	private long auctionID;
 	//auction_event
 	public static final String AUCTION_STARTED="AUCTION_STARTED";
 	public static final String AUCTION_ENDED="AUCTION_ENDED";
 	
-	protected AuctionEvent(String ID, String type, long timestamp, int auctionID) {
+	protected AuctionEvent(String ID, String type, long timestamp, long auctionID) {
 		super(ID, type, timestamp);
 		this.auctionID=auctionID;
 	}
 	
-	public int getAuctionID(){
+	public long getAuctionID(){
 		return auctionID;
 	}
